requirements.txt
Flask
gunicorn
firebase-admin
python-dotenv
stripe
pyparsing<3,>=2.4.2
